# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "XR50TrainingAssetRepo",
    "version": "1.0"
  },
  "paths": {
...
    "/xr50/TrainingAssetRepository/tenant_management": {
      "get": {
        "tags": [
          "tenant_management"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/XR50Tenant"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/XR50Tenant"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/XR50Tenant"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "tenant_management"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/XR50Tenant"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/XR50Tenant"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/XR50Tenant"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/XR50Tenant"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/XR50Tenant"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/XR50Tenant"
                }
              }
            }
          }
        }
      }
    },
    "/xr50/TrainingAssetRepository/tenant_management/{TenantName}": {
      "get": {
        "tags": [
          "tenant_management"
        ],
        "parameters": [
          {
            "name": "TenantName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/XR50Tenant"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/XR50Tenant"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/XR50Tenant"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "tenant_management"
        ],
        "parameters": [
          {
            "name": "TenantName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/xr50/TrainingAssetRepository/tenant_management/{TenantName}/TrainingPrograms": {
      "get": {
        "tags": [
          "tenant_management"
        ],
        "parameters": [
          {
            "name": "TenantName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TrainingProgram"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TrainingProgram"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TrainingProgram"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/xr50/TrainingAssetRepository/tenant_management/{TenantName}/Materials": {
      "get": {
        "tags": [
          "tenant_management"
        ],
        "parameters": [
          {
            "name": "TenantName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Material"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Material"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Material"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/xr50/TrainingAssetRepository/tenant_management/training-management/{TenantName}": {
      "post": {
        "tags": [
          "tenant_management"
        ],
        "parameters": [
          {
            "name": "TenantName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrainingProgram"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/TrainingProgram"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/TrainingProgram"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TrainingProgram"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrainingProgram"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrainingProgram"
                }
              }
            }
          }
        }
      }
    },
    "/xr50/TrainingAssetRepository/tenant_management/material-management/{TenantName}/{ParentMaterialId}": {
      "post": {
        "tags": [
          "tenant_management"
        ],
        "parameters": [
          {
            "name": "TenantName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ParentMaterialId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Material"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Material"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Material"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Material"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Material"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Material"
                }
              }
            }
          }
        }
      }
    },
    "/xr50/TrainingAssetRepository/tenant_management/training-management/{TenantName}/{ProgramName }": {
      "delete": {
        "tags": [
          "tenant_management"
        ],
        "parameters": [
          {
            "name": "TenantName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ProgramName",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ProgramName ",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/xr50/TrainingAssetRepository/tenant_management/material-management/{TenantName}/{ProgramName }/{MateriaId}": {
      "delete": {
        "tags": [
          "tenant_management"
        ],
        "parameters": [
          {
            "name": "TenantName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ProgramName",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "MaterialId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ProgramName ",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "MateriaId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Material": {
        "required": [
          "materialType"
        ],
        "type": "object",
        "properties": {
          "tenantName": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "materialName": {
            "type": "string",
            "nullable": true
          },
          "parentId": {
            "type": "string",
            "nullable": true
          },
          "materialList": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "trainingProgramList": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "materialId": {
            "type": "string",
            "nullable": true
          },
          "materialType": {
            "$ref": "#/components/schemas/MaterialType"
          }
        },
        "additionalProperties": false
      },
      "MaterialType": {
        "enum": [
          0,
          1,
          2,
          3,
          4
        ],
        "type": "integer",
        "format": "int32"
      },
...
      "TrainingProgram": {
        "type": "object",
        "properties": {
          "useCase": {
            "type": "string",
            "nullable": true
          },
          "materialList": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "assetList": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "tenantName": {
            "type": "string",
            "nullable": true
          },
          "programName": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "User": {
        "type": "object",
        "properties": {
          "tenantName": {
            "type": "string",
            "nullable": true
          },
          "fullName": {
            "type": "string",
            "nullable": true
          },
          "userEmail": {
            "type": "string",
            "nullable": true
          },
          "password": {
            "type": "string",
            "nullable": true
          },
          "admin": {
            "type": "boolean"
          },
          "userName": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "XR50Tenant": {
        "type": "object",
        "properties": {
          "tenantName": {
            "type": "string",
            "nullable": true
          },
          "trainingProgramList": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "owncloudGroup": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "owncloudDirectory": {
            "type": "string",
            "nullable": true
          },
          "ownerName": {
            "type": "string",
            "nullable": true
          },
          "owner": {
            "$ref": "#/components/schemas/User"
          },
          "adminList": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}